refactor(app-structure): follow naming convention by swapping models/ and schemas/, renamed files accordingly, and updated the database.py
