+++ app/config/database.py
@@ -1,5 +1,4 @@
 from tortoise import Tortoise
-import ssl
 
 db_url = "sqlite://app/storage/user.sqlite"
 
@@ -7,6 +6,6 @@ db_url = "sqlite://app/storage/user.sqlite"
 async def db_init():
     await Tortoise.init(
         db_url=db_url,
-        modules={"models": ["app.schemas.auth_schemas"]},
+        modules={"models": ["app.models.auth_model"]},
     )
     await Tortoise.generate_schemas()
